staging: media: imx: isi: Disable deprecated drivers
The imx8-mipi-csi2.c and imx8-parallel-csi.c drivers come from the NXP
BSP ISI driver implementation, and are not used by the current ISI
driver. They are kept in the source tree only as convenient reference
during the development of the ISI driver, but don't compile due to a
missing header. Remove the corresponding kernel configuration options to
avoid inadvertent build breakages if the wrong options are selected.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>